Refactored links, added database entries.
Refactored links in index.js to be more readable and more formal, added more database entries(1 table).
Added Category GET/ADD/REMOVE for ticket categories; ADD and REMOVE returning "success" so it can be handled in reactjs as a notification(for example with a toast).
Also added a return to "userExists", fixed it's functionality.